backend/requirements.txt
# Django Core
Django>=5.0,<6.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0

# API Framework
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5
Pillow>=10.2.0

# Authentication
djangorestframework-simplejwt>=5.3.0

# Task Queue & Cache
celery>=5.3.0
redis>=5.0.0
django-redis>=5.4.0

# Rate Limiting
django-ratelimit>=4.1.0

# Utilities
structlog>=24.1.0
django-health-check>=3.18.0
drf-spectacular>=0.27.0
django-environ>=0.11.0
weasyprint>=60.0
reportlab>=4.0

# Production
gunicorn>=21.2.0
uvicorn>=0.27.0
whitenoise>=6.6.0

# Payment providers
stripe>=8.0.0
requests>=2.31.0

# Notifications
twilio>=8.0.0            # optional: WhatsApp/SMS via Twilio
django-celery-beat>=2.5  # Celery periodic tasks (beat scheduler)
